Rename entry file
