+++ app.py
@@ -1,7 +1,7 @@
 """
 	Main file to run the app
 """
-from api import APP
+from api.index import APP
 if __name__ == '__main__':
     # Run the application
     APP.run()
