+++ vercel.json
@@ -1,3 +1,3 @@
 {
-  "rewrites": [{ "source": "/(.*)", "destination": "/api/__init__" }]
+  "rewrites": [{ "source": "/(.*)", "destination": "/api/index" }]
 }
